requirements.txt
pandas >=0.24.0
numpy >=1.16.0
Jinja2 >=2.7
packaging >=16.8
Pillow >=5.2
python-dateutil >=2.1
PyYAML >=3.10
six >=1.5.2
tornado >=4.3
bokeh >=1.2.0
matplotlib >=2.2.0
pytest >=4.1.1